automagical dep finding
