+++ .github/workflows/build_win.yml
@@ -31,7 +31,7 @@ jobs:
       env:
         MKN_CL_PREFERRED: 1
       run: |
-        mkn build -dtKO 2 -g 0 -a "-std:c++17 -EHsc"
-        mkn build -dtOp test_mod -a "-std:c++17 -EHsc"
-        mkn build test pack -Op test -a "-std:c++17 -EHsc"
+        mkn build -dtKO 2 -g 0 -a "-std:c++20 -EHsc"
+        mkn build -dtOp test_mod -a "-std:c++20 -EHsc"
+        mkn build test pack -Op test -a "-std:c++20 -EHsc"
         mkn -p lib_test
