+++ mkn.yaml
@@ -3,7 +3,7 @@
 name: mkn
 version: master
 property:
-  DATE: 11_JAN_2026
+  DATE: 08_MAR_2026
   maiken_location: ${MKN_HOME}/app/mkn/${version}
   maiken_scm: https://github.com/mkn/mkn
   self.deps: mkn.kul
