+++ src/maiken/depmod.cpp
@@ -1,5 +1,5 @@
 /**
-Copyright (c) 2022, Philip Deegan.
+Copyright (c) 2026, Philip Deegan.
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
