+++ src/maiken/globals.cpp
@@ -29,6 +29,7 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 #include "maiken/app.hpp"
+#include "mkn/kul/env.hpp"
 
 namespace maiken {
 std::shared_ptr<AppVars> AppVars::instance;
@@ -37,23 +38,11 @@ std::shared_ptr<AppVars> AppVars::instance;
 maiken::AppVars::AppVars() {
   pks["OS"] = KTOSTRING(__MKN_KUL_OS__);
   pks["HOME"] = mkn::kul::user::home().path();
+  pks["ARCH"] = mkn::kul::env::GET("MKN_ARCH", "x86_64");
   pks["MKN_HOME"] = mkn::kul::user::home(STR_MAIKEN).path();
 
   pks["DATETIME"] = mkn::kul::DateTime::NOW();
   pks["TIMESTAMP"] = std::time(NULL);
-
-  auto const& settings = Settings::INSTANCE();
-
-  if (auto const& local_repo = settings.local_dep_repo())
-    pks["MKN_REPO"] = *local_repo;
-  else
-    pks["MKN_REPO"] = mkn::kul::user::home(mkn::kul::Dir::JOIN(STR_MAIKEN, STR_REPO)).path();
-
-  if (auto const& local_repo = settings.local_mod_repo())
-    pks["MKN_MOD_REPO"] = *local_repo;
-  else
-    pks["MKN_MOD_REPO"] =
-        mkn::kul::user::home(mkn::kul::Dir::JOIN(STR_MAIKEN, STR_MOD_REPO)).path();
 
   std::string ext, pre;
 #if MKN_KUL_IS_WIN
@@ -72,4 +61,20 @@ maiken::AppVars::AppVars() {
   check_set("MKN_OBJ");
   check_set("MKN_LIB_EXT");
   check_set("MKN_LIB_PRE");
+}
+
+void maiken::AppVars::POST_CONSTRUCT() {
+  auto& pks = AppVars::INSTANCE().pks;
+
+  auto const& settings = Settings::INSTANCE();
+  if (auto const& local_repo = settings.local_dep_repo())
+    pks["MKN_REPO"] = *local_repo;
+  else
+    pks["MKN_REPO"] = mkn::kul::user::home(mkn::kul::Dir::JOIN(STR_MAIKEN, STR_REPO)).path();
+
+  if (auto const& local_repo = settings.local_mod_repo())
+    pks["MKN_MOD_REPO"] = *local_repo;
+  else
+    pks["MKN_MOD_REPO"] =
+        mkn::kul::user::home(mkn::kul::Dir::JOIN(STR_MAIKEN, STR_MOD_REPO)).path();
 }
