+++ src/maiken/settings/validator.cpp
@@ -1,0 +1,69 @@
+/**
+Copyright (c) 2026, Philip Deegan.
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are
+met:
+
+    * Redistributions of source code must retain the above copyright
+notice, this list of conditions and the following disclaimer.
+    * Redistributions in binary form must reproduce the above
+copyright notice, this list of conditions and the following disclaimer
+in the documentation and/or other materials provided with the
+distribution.
+    * Neither the name of Philip Deegan nor the names of its
+contributors may be used to endorse or promote products derived from
+this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#include "maiken/settings.hpp"
+#include "maiken/compiler/compilers.hpp"
+
+mkn::kul::yaml::Validator maiken::Settings::validator() const {
+  using namespace mkn::kul::yaml;
+
+  std::vector<NodeValidator> masks;
+  for (auto const& s : Compilers::INSTANCE().keys())
+    masks.push_back(NodeValidator(s, {}, 0, NodeType::STRING));
+
+  return Validator(
+      {NodeValidator("super"), NodeValidator("property", {NodeValidator("*")}, 0, NodeType::MAP),
+       NodeValidator("inc"), NodeValidator("path"),
+       NodeValidator("local",
+                     {NodeValidator("repo"), NodeValidator("mod-repo"), NodeValidator("debugger")},
+                     0, NodeType::MAP),
+       NodeValidator("remote", {NodeValidator("repo"), NodeValidator("mod-repo")}, 0,
+                     NodeType::MAP),
+       NodeValidator("env",
+                     {NodeValidator("name", 1), NodeValidator("value", 1), NodeValidator("mode")},
+                     0, NodeType::NON),
+       NodeValidator("file",
+                     {NodeValidator("type", 1), NodeValidator("compiler", 1),
+                      NodeValidator("linker"), NodeValidator("archiver")},
+                     1, NodeType::LIST),
+#if defined(_MKN_WITH_MKN_RAM_) && defined(_MKN_WITH_IO_CEREAL_)
+       NodeValidator("dist",
+                     {NodeValidator("port"),
+                      NodeValidator("nodes",
+                                    {NodeValidator("host", 1), NodeValidator("port", 1),
+                                     NodeValidator("user"), NodeValidator("pass")},
+                                    0, NodeType::LIST)},
+                     0, NodeType::MAP),
+#endif  // _MKN_WITH_MKN_RAM_ && _MKN_WITH_IO_CEREAL_
+       NodeValidator("compiler", {NodeValidator("mask", masks, 0, NodeType::MAP)}, 0,
+                     NodeType::MAP),
+       NodeValidator("find", {NodeValidator("*")}, 0, NodeType::MAP)});
+}
