+++ src/regex.cpp
@@ -28,9 +28,11 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
+
 #include "maiken/regex.hpp"
+#include "mkn/kul/except.hpp"
+
 #include <regex>
-#include "mkn/kul/log.hpp"
 
 std::vector<std::string> maiken::Regexer::RESOLVE(std::string str) KTHROW(mkn::kul::Exception) {
   std::vector<std::string> v;
