+++ src/settings.cpp
@@ -71,6 +71,8 @@ std::optional<std::string> maiken::Settings::local_mod_repo() const {
 }
 
 maiken::Settings::Settings(std::string const& file_) : mkn::kul::yaml::File(file_) {
+  KLOG(TRC) << file_;
+
   if (root()[STR_LOCAL] && root()[STR_LOCAL][STR_REPO]) {
     mkn::kul::Dir d(root()[STR_LOCAL][STR_REPO].Scalar());
     if (!d.is() && !d.mk())
@@ -103,9 +105,11 @@ maiken::Settings::Settings(std::string const& file_) : mkn::kul::yaml::File(file
     if (f.real() == mkn::kul::File(file()).real())
       KEXCEPT(SettingsException, "super cannot reference itself\n" + file());
     SuperSettings::INSTANCE().cycleCheck(f.real());
-    sup = std::make_unique<Settings>(mkn::kul::yaml::File::CREATE<Settings>(f.full()));
-    for (auto const& p : sup->properties())
-      if (!ps.count(p.first)) ps.insert(p.first, p.second);
+    sup = std::make_unique<Settings>(
+        mkn::kul::yaml::File::CREATE<Settings>(f.full()));  // -> getOrCreate
+    POST_CONSTRUCT(sup.get());
+    for (auto const& [a, b] : sup->properties())
+      if (!ps.count(a)) ps.insert(a, b);
   }
   if (root()[STR_COMPILER] && root()[STR_COMPILER][STR_MASK])
     for (auto const& k : Compilers::INSTANCE().keys())
@@ -113,8 +117,6 @@ maiken::Settings::Settings(std::string const& file_) : mkn::kul::yaml::File(file
         for (auto const& s :
              mkn::kul::String::SPLIT(root()[STR_COMPILER][STR_MASK][k].Scalar(), ' '))
           Compilers::INSTANCE().addMask(s, k);
-
-  resolveProperties();
 }
 
 maiken::Settings& maiken::Settings::INSTANCE() KTHROW(mkn::kul::Exit) {
@@ -159,164 +161,14 @@ bool maiken::Settings::SET(std::string const& s) {
   return 0;
 }
 
-mkn::kul::yaml::Validator maiken::Settings::validator() const {
-  using namespace mkn::kul::yaml;
-
-  std::vector<NodeValidator> masks;
-  for (auto const& s : Compilers::INSTANCE().keys())
-    masks.push_back(NodeValidator(s, {}, 0, NodeType::STRING));
-
-  return Validator(
-      {NodeValidator("super"), NodeValidator("property", {NodeValidator("*")}, 0, NodeType::MAP),
-       NodeValidator("inc"), NodeValidator("path"),
-       NodeValidator("local",
-                     {NodeValidator("repo"), NodeValidator("mod-repo"), NodeValidator("debugger")},
-                     0, NodeType::MAP),
-       NodeValidator("remote", {NodeValidator("repo"), NodeValidator("mod-repo")}, 0,
-                     NodeType::MAP),
-       NodeValidator("env",
-                     {NodeValidator("name", 1), NodeValidator("value", 1), NodeValidator("mode")},
-                     0, NodeType::NON),
-       NodeValidator("file",
-                     {NodeValidator("type", 1), NodeValidator("compiler", 1),
-                      NodeValidator("linker"), NodeValidator("archiver")},
-                     1, NodeType::LIST),
-#if defined(_MKN_WITH_MKN_RAM_) && defined(_MKN_WITH_IO_CEREAL_)
-       NodeValidator("dist",
-                     {NodeValidator("port"),
-                      NodeValidator("nodes",
-                                    {NodeValidator("host", 1), NodeValidator("port", 1),
-                                     NodeValidator("user"), NodeValidator("pass")},
-                                    0, NodeType::LIST)},
-                     0, NodeType::MAP),
-#endif  // _MKN_WITH_MKN_RAM_ && _MKN_WITH_IO_CEREAL_
-       NodeValidator("compiler", {NodeValidator("mask", masks, 0, NodeType::MAP)}, 0,
-                     NodeType::MAP)});
-}
-
-void maiken::Settings::write(mkn::kul::File const& file) KTHROW(mkn::kul::Exit) {
-  mkn::kul::io::Writer w(file);
-  w.write("\n", true);
-
-  bool c = mkn::kul::env::WHICH("clang") || mkn::kul::env::WHICH("clang.exe");
-  bool g = mkn::kul::env::WHICH("gcc") || mkn::kul::env::WHICH("gcc.exe");
-  {
-    auto gcc_pref(mkn::kul::env::GET("MKN_GCC_PREFERRED"));
-    bool use_gcc = gcc_pref.empty() ? 0 : mkn::kul::String::BOOL(gcc_pref);
-    if (c && g && use_gcc) c = 0;
-  }
-
-  w.write("#local:", true);
-  w.write("## Optionnaly override local repository directory", true);
-  w.write("#  repo: <directory>", true);
-  w << mkn::kul::os::EOL();
-
-#ifdef _WIN32
-  auto cl_pref(mkn::kul::env::GET("MKN_CL_PREFERRED"));
-  bool use_cl = cl_pref.empty() ? 0 : mkn::kul::String::BOOL(cl_pref);
-  if (!use_cl && (c || g)) {
-#else
-  if (c || g) {
-#endif
-    w.write("## Add include directories to every compilation", true);
-    w.write("#inc: <directory>", true);
-    w.write("## Add library paths when linking every binary", true);
-    w.write("#path:    <directory>\n", true);
-    w << mkn::kul::os::EOL();
-
-    w.write(
-        "## Modify environement variables for application commands - excludes "
-        "run",
-        true);
-    w.write("#env:", true);
-    w.write("#- name: VAR", true);
-    w.write("#  mode: prepend", true);
-    w.write("#  value: value", true);
-    w << mkn::kul::os::EOL();
-
-    w.write("file:", true);
-    w.write("- type: c:S", true);
-    w.write("  archiver: ar -cr", true);
-    w << "  compiler: " << (c ? "clang" : "gcc") << mkn::kul::os::EOL();
-    w << "  linker: " << (c ? "clang" : "gcc") << mkn::kul::os::EOL();
-    w.write("- type: cpp:cxx:cc", true);
-    w.write("  archiver: ar -cr", true);
-    w << "  compiler: " << (c ? "clan" : "") << "g++" << mkn::kul::os::EOL();
-    w << "  linker: " << (c ? "clan" : "") << "g++" << mkn::kul::os::EOL();
-    w << mkn::kul::os::EOL();
-  }
-
-#ifdef _WIN32
-
-  if (use_cl || (!c && !g)) {
-    auto cl(mkn::kul::env::WHERE("cl.exe"));
-    auto inc(mkn::kul::env::GET("INCLUDE"));
-    auto lib(mkn::kul::env::GET("LIB"));
-
-    if (cl.empty() || inc.empty() || lib.empty()) {
-      w.flush().close();
-      file.rm();
-      KEXIT(1, "gcc or clang not found, vcvars not detected")
-          << mkn::kul::os::EOL()
-          << "\tRun vcvarsall.bat or view mkn wiki to see how to configure "
-             "maiken settings.yaml"
-          << mkn::kul::os::EOL() << "\t@ https://github.com/mkn/mkn/wiki";
-    }
-
-    w.write("inc: ", true);
-    for (auto s : mkn::kul::String::SPLIT(inc, mkn::kul::env::SEP())) {
-      mkn::kul::String::REPLACE_ALL(s, "\\", "/");
-      mkn::kul::String::REPLACE_ALL(s, " ", "\\ ");
-      w << "  " << s << mkn::kul::os::EOL();
-    }
-    w << mkn::kul::os::EOL();
-
-    w.write("path: ", true);
-    for (auto s : mkn::kul::String::SPLIT(lib, mkn::kul::env::SEP())) {
-      mkn::kul::String::REPLACE_ALL(s, "\\", "/");
-      mkn::kul::String::REPLACE_ALL(s, " ", "\\ ");
-      w << "  " << s << mkn::kul::os::EOL();
-    }
-    w << mkn::kul::os::EOL();
-
-    w.write("env:", true);
-    w.write("- name: PATH", true);
-    w.write("  mode: prepend", true);
-    w << "  value: " << mkn::kul::File(cl).dir().real() << mkn::kul::os::EOL();
-    w << mkn::kul::os::EOL();
-
-    w.write("file:", true);
-    w.write("- type: cpp:cxx:cc:c", true);
-    w.write("  archiver: lib", true);
-    w.write("  compiler: cl", true);
-    w.write("  linker: link", true);
-  }
-#endif
-  w << mkn::kul::os::EOL();
-
-  w.write("# Other examples", true);
-  w.write("#- type: cu", true);
-  w.write("#  archiver: ar -cr", true);
-  w.write("#  compiler: nvcc", true);
-  w.write("#  linker: nvcc", true);
-
-  w.write("#- type: m", true);
-  w.write("#  archiver: ar -cr", true);
-  w.write("#  compiler: g++ -lobjc", true);
-  w.write("#  linker: g++", true);
-
-  w.write("#- type: cs", true);
-  w.write("#  compiler: csc", true);
-  w.write("#  linker: csc", true);
-}
-
 mkn::kul::cli::EnvVar maiken::Settings::PARSE_ENV_NODE(YAML::Node const& n,
                                                        Settings const& settings) {
   return maiken::PARSE_ENV_NODE(n, settings, "settings file");
 }
 
-void maiken::Settings::POST_CONSTRUCT() {
-  auto& settings = Settings::INSTANCE();
+void maiken::Settings::POST_CONSTRUCT(Settings* settings_ptr) {
+  settings_ptr = settings_ptr != nullptr ? settings_ptr : &Settings::INSTANCE();
+  auto& settings = *settings_ptr;
   auto& root = settings.root();
 
   if (root[STR_ENV]) {
@@ -334,4 +186,7 @@ void maiken::Settings::POST_CONSTRUCT() {
       }
     }
   }
+
+  settings.resolveProperties();
+  settings.resolveFindables();
 }
